Hommexx/SL: Break main impl file into three.
